Update: Refine the installer setup
