+++ .gitignore
@@ -6,6 +6,7 @@
 !docs
 !debug
 !core
+!deploy
 !device_operation
 !.github
 !develop_tools
@@ -14,22 +15,14 @@
 !.gitignore
 !LICENSE
 !README.md
-!build.bat
 !img.png
-!installer.py
 !main.py
 !minitouch.py
 !requirements.txt
 !ui.png
 !window.py
-!window.spec
-!i18n.pro
 !requirements-i18n.txt
 !requirements-linux.txt
-!dockerfile
-!docker-compose.yaml
-!entrypoint.sh
-src/explore_task_data/__pycache__/normal_task_11.cpython-39.pyc
 *.pyc
 *.xml
 module/.idea/.name
