+++ README.md
@@ -12,9 +12,9 @@ baas 是一款免费开源软件，地址：https://github.com/pur1fying/blue_ar
 
 GUI预览图：
 
-<img src="https://github.com/pur1fying/blue_archive_auto_script/blob/master/ui.png" alt="gui" width="50%">
-<img src="https://github.com/pur1fying/blue_archive_auto_script/blob/master/ui2.png" alt="gui" width="50%">
-<img src="https://github.com/pur1fying/blue_archive_auto_script/blob/master/ui3.png" alt="gui" width="50%">
+<img src="https://github.com/pur1fying/blue_archive_auto_script/blob/master/docs/assets/ui.png" alt="gui" width="50%">
+<img src="https://github.com/pur1fying/blue_archive_auto_script/blob/master/docs/assets/ui2.png" alt="gui" width="50%">
+<img src="https://github.com/pur1fying/blue_archive_auto_script/blob/master/docs/assets/ui3.png" alt="gui" width="50%">
 
 ## 功能 Features
 - **主线**: 自动推图(普通4-26，困难1-26，最新主线<26>已适配)
@@ -53,24 +53,6 @@ GUI预览图：
 - **json：填写自定义的完整地址（如http://127.0.0.1:8081/）**
 - **推送的json格式为:**
 `{"title":"Baas Error","desp":"error..."}`
-
-### Docker 使用方法
-> 本项目拟支持Docker部署，但目前Docker部署仍无法正常运行，如有解决方案，请提交PR
-1. 确保你的电脑中存在Docker, 先clone本项目到本地
-2. 进入项目目录，构建镜像：
-    ```shell
-    docker build -t baas:v1.0 .
-    ```
-3. 运行容器方法一：使用docker run
-    ```shell
-    docker run -it -v "$PWD:/app" -p 5900:5900 baas:v1.0
-    ```
-4. 运行容器方法二：使用docker-compose
-    ```shell
-    docker-compose up -d
-    ```
-5. 可以通过VNC Viewer连接到GUI画面，端口默认为5900
-6. 同时如果你的系统支持显示，可以通过修改`entrypoint.sh`中指示的内容，直接在本地运行。
 
 ## 如何上报bug How to Report Bugs
 在提问题之前至少花费 5 分钟来思考和准备，才会有人花费他的 5 分钟来帮助你。
