+++ deploy/installer/README.md
@@ -1,0 +1,12 @@
+# This Directory contains the installer (Artifact for executable) for the project.
+
+**To use the installer, you need to have the following installed on your system:**
+- **Python 3.6 or higher**
+
+## Maunal
+1. For `Windows` users, please double click on the `build.bat` file. 
+2. For `Linux` users, run the following command in the terminal:
+```bash
+$ bash ./build.sh
+```
+
