+++ deploy/installer/README_CN.md
@@ -1,0 +1,11 @@
+# 该目录包含BAAS的执行文件程序的构建脚本。
+
+**使用安装程序之前，您的系统需要安装以下内容：**
+- **Python 3.6 或更高版本**
+
+## 使用说明
+1. 对于 `Windows` 用户，请双击 `build.bat` 文件。
+2. 对于 `Linux` 用户，请在终端中运行以下命令：
+```bash
+$ bash ./build.sh
+```
