+++ deploy/installer/requirements.installer.txt
@@ -1,0 +1,4 @@
+tqdm
+dulwich
+requests
+pyinstaller
